requirements.txt
amqp==2.6.0
asgiref==3.2.10
Babel==2.8.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
bleach==3.1.5
celery==4.4.6
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
clamd==1.0.2
cryptography==2.9.2
cssselect==1.1.0
defusedxml==0.7.0rc1
Django==2.2.14
django-appconf==1.0.4
django-bootstrap-breadcrumbs==0.9.2
django-ckeditor==5.9.0
django-cleanup==5.0.0
django-crispy-forms==1.9.2
django-cryptography==1.0
django-js-asset==1.2.2
django-resized==0.3.11
django-reversion==3.0.7
django-safe-filefield==0.3.1
django-social-share==1.4.0
django-taggit==1.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
future==0.18.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.7.0
is-safe-url==1.0
Jinja2==2.11.2
kombu==4.6.11
ldif3==3.1.1
lxml==4.5.1
Markdown==3.2.2
markdown2==2.3.9
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pyquery==1.4.1
python-card-me==0.9.3
python-dateutil==2.8.1
python-magic==0.4.18
python-magic-bin==0.4.14
python3-openid==3.1.0
pytz==2020.1
readtime==1.1.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-core==3.3.3
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.1
surlex==0.2.0
unicodecsv==0.14.1
urllib3==1.25.9
vine==1.3.0
webencodings==0.5.1
zipp==3.1.0
